Promote Certificate#verify to a top level api in android sdk
