+++ android/trifle/build.gradle.kts
@@ -40,6 +40,7 @@ dependencies {
   androidTestImplementation("androidx.test.ext:junit:1.1.5")
   androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
   androidTestImplementation("junit:junit:4.13.2")
+  androidTestImplementation(project(":jvm-testing"))
 
   api(project(":jvm"))
 }
