+++ jvm/src/main/kotlin/app/cash/trifle/Certificate.kt
@@ -1,11 +1,7 @@
 package app.cash.trifle
 
-import app.cash.trifle.CertificateRequest.PKCS10Request
-import app.cash.trifle.TrifleErrors.CSRMismatch
-import app.cash.trifle.internal.validators.CertChainValidatorFactory
 import okio.ByteString.Companion.toByteString
 import org.bouncycastle.cert.X509CertificateHolder
-import java.util.Date
 import app.cash.trifle.protos.api.alpha.Certificate as CertificateProto
 
 /**
@@ -28,48 +24,28 @@ data class Certificate internal constructor(
     ).encode()
 
   /**
-   * Verify that the provided certificate matches what we expected.
-   * It matches the CSR that we have and the root cert is what
-   * we expect.
+   * Verify that the provided certificate matches the CSR that we have.
    *
    * @param certificateRequest -  request used to generate this certificate
-   * @param ancestorCertificateChain - list of certificates preceding *this* one.  Namely, the first
-   *   entry should be the certificate corresponding to the issuer of this certificate, and each
-   *   thereafter should be the issuer of the one before it.
-   * @param anchorCertificate - the trust anchor against which we would like to verify the
-   *   ancestorCertificateChain. This may be the terminal (root) certificate of the chain or may be
-   *   an intermediate certificate in the chain which is already trusted.
-   * @param date - The date to use for verification against certificates' validity windows. If null,
-   *   the current time is used.
    *
    * @return - [Result] indicating [Result.isSuccess] or [Result.isFailure]:
    * - success value is expressed as a [Unit] (Nothing)
-   * - failure value is expressed as a [TrifleErrors]
+   * - failure value is expressed as a [TrifleErrors.CSRMismatch] if attributes are mismatched
    */
-  fun verify(
-    certificateRequest: CertificateRequest,
-    ancestorCertificateChain: List<Certificate>,
-    anchorCertificate: Certificate,
-    date: Date? = null
-  ): Result<Unit> {
-    // First check to see if the certificate chain validates
-    val certChainResult = CertChainValidatorFactory.get(anchorCertificate, date)
-      .validate(listOf(this) + ancestorCertificateChain)
-
-    return certChainResult.mapCatching {
-      val x509Certificate = X509CertificateHolder(certificate)
-      when (certificateRequest) {
-        is PKCS10Request -> {
-          // Certificate chain matches, check with certificate request.
-          // TODO(dcashman): Check other attributes as well.
-          if (certificateRequest.pkcs10Req.subject != x509Certificate.subject ||
-            certificateRequest.pkcs10Req.subjectPublicKeyInfo != x509Certificate.subjectPublicKeyInfo
-          ) {
-            throw CSRMismatch
-          }
+  fun verify(certificateRequest: CertificateRequest): Result<Unit> {
+    val x509Certificate = X509CertificateHolder(certificate)
+    when (certificateRequest) {
+      is CertificateRequest.PKCS10Request -> {
+        // Certificate chain matches, check with certificate request.
+        // TODO(dcashman): Check other attributes as well.
+        if (certificateRequest.pkcs10Req.subject != x509Certificate.subject ||
+          certificateRequest.pkcs10Req.subjectPublicKeyInfo != x509Certificate.subjectPublicKeyInfo
+        ) {
+          return Result.failure(TrifleErrors.CSRMismatch)
         }
       }
     }
+    return Result.success(Unit)
   }
 
   override fun equals(other: Any?): Boolean {
